Int and Float list nodes.
rarely useful, but it is getting confusing to rely on different extensions for different types.
I am not aware of any extension that allows to fetch multiple items from a list like my FromListGetXXXX. Maybe something similar already exists?

Maybe should just bypass like I saw in some other extension for other data types?
Or fork that extension and implement the fetching of multiple individual items in that extension?

My approach in AnyToAny node was to not bypass type check and force the user to use a modified version of comfy; so, to be consistent, I should not bypass the check in my extension with a generic node, and should fork someone else's extension that uses such approach.

Will postpone any generic implementation for now...